requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
